nocaj

+++ include/engine/camera.hpp
@@ -64,6 +64,6 @@ public:
     glm::mat4 get_projection_matrix();
     glm::mat4 get_view_projection_matrix();
 
-    glm::vec2 screen_to_world(float screen_x, float screen_y);
+    glm::ivec2 screen_to_world(float screen_x, float screen_y);
     glm::vec2 world_to_screen(float world_x, float world_y);
 };
